pyproject.toml
[project]
name = "ai-architect-rag-srt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.12"
dependencies = [
    "altair==4.2.2",
    "anthropic>=0.86.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.5",
    "datasets==3.6.0",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.6",
    "gradio>=6.10.0",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "jupyter-dash>=0.4.2",
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "matplotlib>=3.10.8",
    "modal>=1.4.0",
    "nbformat>=5.10.4",
    "numpy>=2.4.3",
    "ollama>=0.6.1",
    "openai>=2.30.0",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "protobuf==3.20.2",
    "psutil>=7.2.2",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "sentence-transformers>=5.3.0",
    "setuptools>=82.0.1",
    "speedtest-cli>=2.1.3",
    "streamlit>=1.55.0",
    "streamlit-authenticator>=0.4.2",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
    "wandb>=0.25.0",
]
